:bricks: rename lib to utils
