+++ src/actions/Pages/getExecPage.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { getPayload } from "@/libs/payload";
+import { getPayload } from "@/utils/payload";
 
 export const getExecPage = async () => {
     const payload = await getPayload();
