+++ src/actions/getAllEvents.ts
@@ -1,7 +1,7 @@
 "use server";
 
-import { SLUG } from "@/libs/enums/slug";
-import { getPayload } from "@/libs/payload";
+import { SLUG } from "@/utils/enums/slug";
+import { getPayload } from "@/utils/payload";
 
 // return all events
 export const getAllEvents = async () => {
