+++ src/actions/getProject.ts
@@ -2,8 +2,8 @@
 
 import { Project } from "@/types/payload-types";
 
-import { SLUG } from "@/libs/enums/slug";
-import { getPayload } from "@/libs/payload";
+import { SLUG } from "@/utils/enums/slug";
+import { getPayload } from "@/utils/payload";
 
 // Return single project by year and slug
 export const getProject = async (year: string, slug: string): Promise<Project | undefined> => {
