+++ src/app/(site)/(landing)/_components/Intro.tsx
@@ -4,7 +4,7 @@ import { ReactNode } from "react";
 
 import { Variants, motion } from "motion/react";
 
-import { easeOutQuad } from "@/libs/ease";
+import { easeOutQuad } from "@/utils/ease";
 
 type IntroProps = {
     children: ReactNode;
