+++ src/app/(site)/about/faq/_components/FaqButton.tsx
@@ -1,5 +1,5 @@
 import Plus from "@/assets/svg/Plus";
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 type FaqButtonProps = {
     rotate?: boolean;
