+++ src/app/(site)/about/faq/_components/FaqSelector.tsx
@@ -8,7 +8,7 @@ import { Faq } from "@/types/models";
 import { FaqSection } from "@/types/pages/FaqPage";
 
 import ChevronDown from "@/assets/svg/ChevronDown";
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 import { FaqItem } from "./FaqItem";
 import { FaqTab } from "./FaqTab";
