+++ src/app/(site)/projects/[...fullSlug]/_component/IndividualProject.tsx
@@ -2,7 +2,7 @@ import { Project } from "@/types/models";
 
 import { getTechnologySvgWithName } from "@/assets/svg/technologies";
 import Duo from "@/components/Duo";
-import { toTitleCase } from "@/libs/utils";
+import { toTitleCase } from "@/utils/misc";
 
 import TeamMemberCard from "./TeamMemberCard";
