+++ src/app/(site)/projects/_components/FilterTag.tsx
@@ -1,6 +1,4 @@
-import React from "react";
-
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 type FilterTagProps = {
     year: string;
