+++ src/assets/svg/technologies/index.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, SVGAttributes } from "react";
 
-import { hasKey } from "@/libs/utils";
+import { hasKey } from "@/utils/misc";
 
 import { AWS } from "./AWS";
 import { Astro } from "./Astro";
