+++ src/collections/Events.ts
@@ -2,7 +2,7 @@ import type { CollectionConfig } from "payload";
 
 import { EventCategories } from "@/types/models";
 
-import { SLUG } from "@/libs/enums/slug";
+import { SLUG } from "@/utils/enums/slug";
 
 export const Events: CollectionConfig = {
     slug: SLUG.EVENTS,
