+++ src/collections/TestSlug.ts
@@ -1,7 +1,7 @@
 import type { CollectionConfig } from "payload";
 
 import { technologiesList } from "@/assets/svg/technologies";
-import { SLUG } from "@/libs/enums/slug";
+import { SLUG } from "@/utils/enums/slug";
 
 export const TestSlug: CollectionConfig = {
     slug: SLUG.TEST,
