+++ src/components/Draw.tsx
@@ -4,7 +4,7 @@ import React from "react";
 
 import { motion } from "motion/react";
 
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 type DrawProps = {
     children: React.ReactNode;
