+++ src/components/ImageCover.tsx
@@ -1,9 +1,7 @@
-import React from "react";
-
 import { StaticImport } from "next/dist/shared/lib/get-img-props";
 import Image, { StaticImageData } from "next/image";
 
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 export type ImageFitProps = {
     src: string | StaticImageData | StaticImport;
