+++ src/components/article/TableOfContents.tsx
@@ -6,7 +6,7 @@ import { VariantProps, tv } from "tailwind-variants";
 
 import SimpleLoading from "@/components/loading/SimpleLoading";
 import { useScroll } from "@/components/scroll/ScrollContext";
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 type ContentlistType = { level: number; label: string; id: string }[];
