+++ src/components/layout/ThreePageLayout.tsx
@@ -7,9 +7,9 @@ import dynamic from "next/dynamic";
 import { useProgress } from "@react-three/drei";
 import { motion } from "motion/react";
 
-import { fadeopacity } from "@/libs/animations";
 import { useReduceMotion } from "@/providers/ReduceMotionProvider";
 import { useWebGL } from "@/providers/WebGLProvider";
+import { fadeopacity } from "@/utils/animations";
 
 import ReduceMotionButton from "../ReduceMotionButton";
 import NavigationBar from "../navigation/navbar/NavigationBar";
