+++ src/components/navigation/UnderlineLink.tsx
@@ -2,7 +2,7 @@ import { AnchorHTMLAttributes, ReactNode } from "react";
 
 import Link, { LinkProps } from "next/link";
 
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 // Need to omit href from AnchorHTMLAttributes because it conflicts with Next LinkProps
 type UnderlineLinkProps = {
