+++ src/components/navigation/footer/FooterColumn.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 import { UnderlineLink } from "../UnderlineLink";
 import type { FooterColumn } from "./_data/footerTypes";
