+++ src/components/navigation/footer/FooterSocials.tsx
@@ -4,7 +4,7 @@ import { HTMLAttributes } from "react";
 
 import { Anchor } from "@/components/primitives/Anchor";
 import { useScroll } from "@/components/scroll/ScrollContext";
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 import { Button } from "../../primitives/Button";
 import { FooterSocial } from "./_data/footerTypes";
