+++ src/components/primitives/Button.tsx
@@ -6,7 +6,7 @@ import Link, { LinkProps } from "next/link";
 
 import { type VariantProps, tv } from "tailwind-variants";
 
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 
 const button = tv({
     base: "flex items-center justify-center gap-2 rounded-full px-6 py-2 font-bold whitespace-nowrap transition duration-200 hover:cursor-pointer",
