+++ src/components/text/AnimatedText.tsx
@@ -1,12 +1,11 @@
 "use client";
 
-import React from "react";
 
 import { Variants, motion } from "motion/react";
 
 import { HTMLAtrributeNoMotion } from "@/types/motion";
 
-import { cn } from "@/libs/utils";
+import { cn } from "@/utils/misc";
 import { textVariantDefault } from "@/variants/animatedTextVariants";
 
 type TextProps = {
